+++ examples/modules/vellum_demo/migrations/0_1_to_0_2.py
@@ -1,6 +1,0 @@
-"""Add vellum.demo.comment for Vellum relation tests."""
-
-
-def upgrade(env):
-    env["vellum.demo.comment"]._setup_table(env.conn)
-    env["vellum.demo.note"]._setup_table(env.conn)